+++ backend/data-access/research_model.ts
@@ -1,0 +1,91 @@
+import { PrismaClient } from '@prisma/client'
+
+
+
+export async function getResearchNotes(projectID: string) {
+    const prisma = new PrismaClient()
+    try {
+
+        const notes = await prisma.researchNote.findMany(
+            {
+                where: {
+                    id: projectID,
+                },
+                include: {
+                    link: true
+                }
+            }
+        )
+
+        return notes;
+    } catch (err: any) {
+        console.log(err)
+        return null;
+    } finally {
+        prisma.$disconnect()
+    }
+
+}
+
+export async function createResearchNote(title: string, details: string, userID: number, sprint: string,urlList:any) {
+    const prisma = new PrismaClient()
+    try {
+
+        const notes = await prisma.researchNote.create(
+            {
+                data:{
+                    sprintID:sprint,
+                    title:title,
+                    userID:userID,
+                    details:details,
+                    link:{
+                        createMany:{
+                            data:urlList
+                        },
+                    }
+                },
+                
+            }
+        )
+
+        return notes;
+    } catch (err: any) {
+        console.log(err)
+        return null;
+    } finally {
+        prisma.$disconnect()
+    }
+}
+
+export async function updateResearchNote(noteID: string,title: string, details: string, userID: number, sprint: string,urlList:any) {
+    const prisma = new PrismaClient()
+    try {
+
+        const notes = await prisma.researchNote.update(
+            {
+                where:{
+                    id:noteID
+                },
+                data:{
+                    sprintID:sprint,
+                    title:title,
+                    userID:userID,
+                    details:details,
+                    link:{
+                        createMany:{
+                            data:urlList
+                        },
+                    }
+                },
+                
+            }
+        )
+
+        return notes;
+    } catch (err: any) {
+        console.log(err)
+        return null;
+    } finally {
+        prisma.$disconnect()
+    }
+}
